Adds slug field to RouterURL
slug is URL encoded version of pathname + search.
It can be used for all data fetching and caching
to avoid having to build pathname + search everywhere on the fly.

It is url encoded so that it can be used in url for data fetching:
/_aleph/data/{slug}.json

Also fixes bug in Anchor.isCurrent being incorrect when adding search:
/foo -> /foo?bar=1